refactor: RandomUtil 변경 및 StringUtil 파라미터 추가
