+++ src/main/java/racingcar/util/StringUtil.java
@@ -9,9 +9,9 @@ public class StringUtil {
     private StringUtil() {
     }
 
-    public static List<String> split(String values) {
+    public static List<String> split(String values, String Delimiter) {
         try {
-            return Arrays.stream(values.split(",")).collect(Collectors.toList());
+            return Arrays.stream(values.split(Delimiter)).collect(Collectors.toList());
         } catch (IllegalArgumentException illegalArgumentException) {
             throw new IllegalArgumentException("잘못된 입력 값 입니다.");
         }
