+++ src/test/java/racingcar/RandomNumberUtilTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
-import racingcar.view.RandomNumberUtil;
+import racingcar.util.RandomNumberUtil;
 
 import static org.assertj.core.api.Assertions.assertThat;
